# processor/src/main/resources/swagger/foodbox.yaml
openapi: 3.0.1
info:
  title: API по питанию в образовательных учреждениях Москвы
  description: API по питанию в образовательных учреждениях Москвы
  version: 1.0.1
servers:
  - url: http://mes-api.mos.ru/meals
  - url: http://mes-api-test.mos.ru/meals
security:
  - apiKey: []
    agent: []
  - bearerAuth: []
paths:
...
  /orders/foodbox:
    parameters:
      - $ref: '#/components/parameters/clientId'
    post:
      tags:
        - Фудбокс-заказ
      summary: Создаёт заказ
      operationId: addPersonFoodboxOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoodboxOrder'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderErrorInfo'
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentFoodboxOrderInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Фудбокс-заказ
      summary: Удаляет заказ.
      operationId: deletePersonFoodboxOrder
      parameters:
        - name: id
          description: Идентификатор заказа.
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Фудбокс-заказ
      summary: Возвращает список всех фудбокс-заказов.
      operationId: getPersonFoodboxOrderList
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/foodboxOrderNumber'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryFoodboxOrderInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    clientId:
      name: clientId
      description: Идентификатор персоны из МЭШ.Контингент
      in: query
      required: true
      style: form
      explode: true
      schema:
        $ref: '#/components/schemas/clientId'
...
    From:
      name: from
      description: Дата начала выборки.
      in: query
      required: true
      schema:
        type: string
        format: date
    To:
      name: to
      description: Дата конца выборки.
      in: query
      required: true
      schema:
        type: string
        format: date
...
    Sort:
      name: sort
      description: Сортировка по дате факта.
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    foodboxOrderNumber:
      name: foodboxOrderNumber
      description: Номер фудбокс-заказа
      in: query
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Не хватает обязательных параметров, неправильное значение парметров или другая ошибка, связанная с формулировкой запроса.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Для доступа к запрашиваемому ресурсу требуется аутентификация.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Клиент не уполномочен совершать операции с запрошенным ресурсом.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Не найден запрашиваемый ресурс.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Внутренняя ошибка сервера, сбой.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      description: Ответ с ошибкой.
      required:
        - code
        - description
      type: object
      properties:
        code:
          minLength: 1
          type: string
          description: Идентификатор ошибки.
        description:
          minLength: 1
          type: string
          description: Описание ошибки.
        details:
          type: object
          properties: {}
          description: Детализация ошибки.
...
    OrderDish:
      description: Блюдо в заказе.
      required:
        - id
        - dishId
        - amount
        - price
        - buffetCategoriesId
        - buffetCategoriesName
      type: object
      properties:
        dishId:
          type: integer
          description: Идентификатор блюда.
          format: int64
        amount:
          type: integer
          description: Заказываемое количество блюда.
        name:
          type: string
          description: Название блюда.
        price:
          type: integer
          description: Цена блюда в копейках.
          format: int64
        buffetCategoriesId:
          type: integer
          description: Идентификатор категории.
          format: int64
        buffetCategoriesName:
          type: string
          description: Название категории.
    FoodboxOrder:
      description: Cоздание фудбокс-заказа.
      type: object
      required:
        - id
        - createdAt
        - foodboxAllowed
      properties:
        id:
          type: integer
          description: Идентификатор заказа.
          format: int64
        createdAt:
          type: string
          description: Дата и время совершения заказа.
          format: date-time
        dishes:
          description: Список блюд в заказе.
          type: array
          items:
            $ref: '#/components/schemas/OrderDish'
        orderPrice:
          type: integer
          description: Общая стоимость заказа в копейках
          format: int64
    FoodboxOrderInfo:
      description: Передача данных о заказе (в контексте списка).
      type: object
      required:
        - id
        - status
        - dishes
      properties:
        id:
          type: integer
          description: Идентификатор заказа.
          format: int64
        status:
          type: string
          description: Статус заказа в системе. При запросе списка заказов передавать статусы всех заказов, в т.ч. и текущих
        expiresAt:
          type: string
          description: Дата и время, до которого клиент может забрать заказ
          format: date-time
        timeOrder:
          type: string
          description: Дата и время создания заказа
          format: date-time
        issueTime:
          type: string
          description: Дата и время получения заказа. Параметр обязателен для выданных заказов
          format: date-time
        foodboxOrderNumber:
          type: integer
          description: Номер фудбокс-заказа
          format: int64
        dishes:
          type: array
          items:
            $ref: '#/components/schemas/OrderDish'
        orderPrice:
          type: integer
          description: Общая стоимость заказа в копейках
          format: int64
    HistoryFoodboxOrderInfo:
      description: История заказов или получение данных по конкретному заказу фудбокса
      type: object
      properties:
        foodboxAvailabilityForEO:
          type: boolean
          description: Признак доступности использования фудбокса для образовательной организации
        foodboxAvailability:
          type: boolean
          description: Признак доступности использования фудбокса
        ordersInfo:
          type: array
          items:
            $ref: '#/components/schemas/FoodboxOrderInfo'
        ordersAmount:
          type: integer
          description: Количество заказов в списке
          format: int64
...
    CurrentFoodboxOrderInfo:
      description: Передача данных о текущем заказе
      type: object
      required:
        - id
        - foodboxOrderNumber
        - status
        - dishes
        - expiresAt
        - timeOrder
        - orderPrice
      properties:
        id:
          type: integer
          description: Идентификатор заказа.
          format: int64
        foodboxOrderNumber:
          type: integer
          description: Номер фудбокс-заказа
          format: int64
        status:
          type: string
          description: Статус заказа в системе
        dishes:
          type: array
          items:
            $ref: '#/components/schemas/OrderDish'
        expiresAt:
          type: string
          description: Дата и время, до которого клиент может забрать заказ
          format: date-time
        timeOrder:
          type: string
          description: Дата и время создания заказа
          format: date-time
        orderPrice:
          type: integer
          description: Общая стоимость заказа в копейках
          format: int64
    OrderErrorInfo:
      description: Информация о балансе ЛС клиента
      type: object
      required:
        - code
        - information
      properties:
        code:
          type: integer
          format: int64
          description: Код ошибки, возвращаемый в ответе
          example: 10
        information:
          type: string
          description: Информация возвращаемая в ответе
          example: Заказ не попадает в рамки времени работы буфета
        currentBalanceLimit:
          type: integer
          format: int64
          description: Лимит дневных трат. Атрибут передаётся в случае ошибки, связанной с превышением дневного лимита
        currentBalance:
          type: integer
          format: int64
          description: Остаток денежных средств. Атрибут передаётся в случае ошибки, связанной с нехваткой средств на балансе при попытке заказа
...
    clientId:
      description: Идентификатор персоны.
      type: object
      properties:
        contractId:
          type: integer
          description: Номер лицевого счета.
          format: int64
        staffId:
          description: Идентификатор сотрудника из реестра кадров.
          type: integer
          format: int64
        personId:
          description: Идентификатор персоны из МЭШ.Контингента.
          type: string
          format: uuid
...
  securitySchemes:
    apiKey:
      type: apiKey
      name: X-Api-Key
      in: header
      description: Выполнение операций на основе clientId
    agent:
      type: apiKey
      name: agent
      in: header
      description: Информация по агенту, от имени которого выполняется операция { "stf":12, "ctg":"07c6782a-edf6-461f-a8b0-0dfc3df789d1"}
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Вызовы на основе токена АУПД (personId и staffId)