# processor/src/web/java/ru/axetta/ecafe/processor/web/partner/meals/models/OrderDish.java
package ru.axetta.ecafe.processor.web.partner.meals.models;

import java.util.Objects;

/**
 * Блюдо в заказе.
 */
public class OrderDish {
    private Long dishId = null;
    private Integer amount = null;
    private String name = null;
    private Long price = null;
    private Long buffetCategoriesId = null;
    private String buffetCategoriesName = null;
    public OrderDish dishId(Long dishId) {
        this.dishId = dishId;
        return this;
    }



    /**
     * Идентификатор блюда.
     * @return dishId
     **/
    public Long getDishId() {
        return dishId;
    }
    public void setDishId(Long dishId) {
        this.dishId = dishId;
    }
    public OrderDish amount(Integer amount) {
        this.amount = amount;
        return this;
    }



    /**
     * Заказываемое количество блюда.
     * @return amount
     **/
    public Integer getAmount() {
        return amount;
    }
    public void setAmount(Integer amount) {
        this.amount = amount;
    }
    public OrderDish name(String name) {
        this.name = name;
        return this;
    }



    /**
     * Название блюда.
     * @return name
     **/
    public String getName() {
        return name;
    }
    public void setName(String name) {
        this.name = name;
    }
    public OrderDish price(Long price) {
        this.price = price;
        return this;
    }



    /**
     * Цена блюда в копейках.
     * @return price
     **/
    public Long getPrice() {
        return price;
    }
    public void setPrice(Long price) {
        this.price = price;
    }
    public OrderDish buffetCategoriesId(Long buffetCategoriesId) {
        this.buffetCategoriesId = buffetCategoriesId;
        return this;
    }



    /**
     * Идентификатор категории.
     * @return buffetCategoriesId
     **/
    public Long getBuffetCategoriesId() {
        return buffetCategoriesId;
    }
    public void setBuffetCategoriesId(Long buffetCategoriesId) {
        this.buffetCategoriesId = buffetCategoriesId;
    }
    public OrderDish buffetCategoriesName(String buffetCategoriesName) {
        this.buffetCategoriesName = buffetCategoriesName;
        return this;
    }



    /**
     * Название категории.
     * @return buffetCategoriesName
     **/
    public String getBuffetCategoriesName() {
        return buffetCategoriesName;
    }
    public void setBuffetCategoriesName(String buffetCategoriesName) {
        this.buffetCategoriesName = buffetCategoriesName;
    }
    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        OrderDish orderDish = (OrderDish) o;
        return Objects.equals(this.dishId, orderDish.dishId) &&
                Objects.equals(this.amount, orderDish.amount) &&
                Objects.equals(this.name, orderDish.name) &&
                Objects.equals(this.price, orderDish.price) &&
                Objects.equals(this.buffetCategoriesId, orderDish.buffetCategoriesId) &&
                Objects.equals(this.buffetCategoriesName, orderDish.buffetCategoriesName);
    }

    @Override
    public int hashCode() {
        return Objects.hash(dishId, amount, name, price, buffetCategoriesId, buffetCategoriesName);
    }

    @Override
    public String toString() {
        StringBuilder sb = new StringBuilder();
        sb.append("class OrderDish {\n");

        sb.append("    dishId: ").append(toIndentedString(dishId)).append("\n");
        sb.append("    amount: ").append(toIndentedString(amount)).append("\n");
        sb.append("    name: ").append(toIndentedString(name)).append("\n");
        sb.append("    price: ").append(toIndentedString(price)).append("\n");
        sb.append("    buffetCategoriesId: ").append(toIndentedString(buffetCategoriesId)).append("\n");
        sb.append("    buffetCategoriesName: ").append(toIndentedString(buffetCategoriesName)).append("\n");
        sb.append("}");
        return sb.toString();
    }

    /**
     * Convert the given object to string with each line indented by 4 spaces
     * (except the first line).
     */
    private String toIndentedString(Object o) {
        if (o == null) {
            return "null";
        }
        return o.toString().replace("\n", "\n    ");
    }

}